requirements.txt
absl-py==0.2.1
astor==0.6.2
bleach==1.5.0
boto==2.48.0
boto3==1.7.24
botocore==1.10.24
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
docutils==0.14
gast==0.2.0
gensim==3.4.0
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
jmespath==0.9.3
Keras==2.1.6
Markdown==2.6.11
numpy==1.14.3
pandas==0.22.0
protobuf==3.5.2.post1
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.5.7
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
matplotlib==2.2.2
